Tier B Phase 3d-ii (values): move extractMethod + stableStringifyJson to the SDK
Relocate the two pure helpers the widget renderer needs at module scope into
@mcpjam/sdk/widget-runtime (framework-free), so the renderer (relocating in the
cutover slice) and the inspector share one impl rather than duplicating across
the package boundary.

- New sdk/src/widget-runtime/json-utils.ts: extractMethod (JSON-RPC / OpenAI-shim
  method label) + stableStringifyJson (+ its private canonicalizer). Exported
  from the widget-runtime barrel.
- Inspector @/stores/traffic-log-store + @/lib/client-config import them from the
  SDK (for their own internal use) and re-export for back-compat; all existing
  import sites unchanged.

Verified: SDK build/typecheck/test (1492); inspector typecheck:client + traffic-
log + renderer tests; root typecheck. No behavior change.
